댓글 저장부분 컴퍼넌트화 시킴
