+++ src/main/front/src/data/mock_feed.json
@@ -1,6 +1,6 @@
 [
     {
-      "id":1,
-      "text":"hello"
+      "email":"test@test.com",
+      "context":"hello"
     }
 ]
